# internal/api/openapi.yml
openapi: 3.0.3
info:
  title: DBBat API
  description: |
    REST API for DBBat PostgreSQL Observability Proxy.

    DBBat is a transparent PostgreSQL proxy for query observability, access control, and safety.

    ## Authentication

    The API supports two authentication methods:

    - **Basic Auth**: HTTP Basic Authentication using username and password
    - **Bearer Token**: API key authentication using `Authorization: Bearer <api-key>`

    API keys cannot create or revoke other API keys (security restriction) - these operations require Basic Auth.

    ## Roles

    Users have one or more roles that determine their access:

    - **admin**: Full access to all resources and operations
    - **viewer**: Read-only access to observability data (connections, queries, audit)
    - **connector**: Can only access databases they have active grants for

    ## Password Change Requirement

    Users must change their initial password before logging in.
    Login attempts from users who haven't changed their password will receive a `403 Forbidden` response with error code `password_change_required`.
    Users must then call `PUT /auth/password` with their username and credentials to change their password before they can log in.

    ## Rate Limiting

    All authenticated endpoints are rate-limited. Response headers include:

    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in current window
...

...

...

    - 1-2 failures: no delay
...

...
servers:
...
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user and session info
      description: |
        Returns information about the currently authenticated user and their session.

        Use this endpoint to:
        - Validate that a session token is still valid
        - Check if password change is required
        - Get session expiration time
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user and session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/{uid}:
    parameters:
      - $ref: '#/components/parameters/UserUID'

    get:
      tags:
        - Users
      summary: Get user by UID
      description: Retrieves a specific user by their UID.
      operationId: getUser
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    put:
      tags:
        - Users
      summary: Update user
      description: |
        Updates a user account.

        - Non-admins can only update their own password
        - Non-admins cannot change roles
        - API keys cannot change passwords (requires Basic Auth)
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: |
        Deletes a user account. Requires admin role.

        Note: Cannot delete your own user account.
      operationId: deleteUser
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: API key ID
        schema:
          type: string
          format: uuid

    get:
      tags:
        - API Keys
      summary: Get API key by ID
      description: |
        Retrieves a specific API key.

        Non-admins can only see their own keys.
      operationId: getAPIKey
      responses:
        '200':
          description: API key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      tags:
        - API Keys
      summary: Revoke API key
      description: |
        Revokes an API key.

        - Non-admins can only revoke their own keys
        - Requires Web Session or Basic Auth (API keys cannot revoke API keys)
      operationId: revokeAPIKey
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        '204':
          description: API key revoked successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication (username:password)
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication (Bearer token)
  parameters:
    UserUID:
      name: uid
      in: path
      required: true
      description: User UID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    MeResponse:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: User UID
        username:
          type: string
          description: Username
        roles:
          type: array
          items:
            type: string
            enum: [admin, viewer, connector]
          description: User roles
        password_change_required:
          type: boolean
          description: Whether user must change their password
        session:
          $ref: '#/components/schemas/SessionInfo'
      required:
        - uid
        - username
        - roles
        - password_change_required
        - session
    SessionInfo:
      type: object
      description: Current session information
      properties:
        expires_at:
          type: string
          format: date-time
          description: Session expiration time
        created_at:
          type: string
          format: date-time
          description: Session creation time
...
    User:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier
        username:
          type: string
          description: Username
        roles:
          type: array
          items:
            type: string
            enum: [admin, viewer, connector]
          description: User roles
        rate_limit_exempt:
          type: boolean
          description: Whether user bypasses rate limiting
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - uid
        - username
        - roles
        - rate_limit_exempt
        - created_at
        - updated_at
...
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
          description: New password
        roles:
          type: array
          items:
            type: string
            enum: [admin, viewer, connector]
          description: New roles (admin only)

    # Database schemas
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
        user_id:
          type: string
          format: uuid
          description: User UID
        name:
          type: string
          description: Key name
        key_prefix:
          type: string
          description: Key prefix for identification
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: When the key expires
        last_used_at:
          type: string
          format: date-time
          nullable: true
          description: When the key was last used
        request_count:
          type: integer
          format: int64
          description: Total requests made with this key
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        revoked_at:
          type: string
          format: date-time
          nullable: true
          description: When the key was revoked
        revoked_by:
          type: string
          format: uuid
          nullable: true
          description: User who revoked the key
      required:
        - id
        - user_id
        - name
        - key_prefix
        - request_count
        - created_at
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: rate_limit_exceeded
        message:
          type: string
          example: Too many requests. Please retry after 60 seconds.
        retry_after:
          type: integer
          description: Seconds until retry is allowed
      required:
        - error
        - message
        - retry_after
    PasswordChangeRequiredError:
      type: object
      description: Returned when user hasn't changed their initial password
      properties:
        error:
          type: string
          enum: [password_change_required]
          example: password_change_required
        message:
          type: string
          example: You must change your password before accessing the API
      required:
        - error
        - message
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
      required:
        - message
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Insufficient permissions or password change required
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Error'
              - $ref: '#/components/schemas/PasswordChangeRequiredError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until retry is allowed
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests per minute
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'